fix satoken interceptor
