+++ quarkus-satoken-main/quarkus-satoken-resteasy/integration-tests/src/main/resources/application.properties
@@ -1,0 +1,1 @@
+sa-token.annotation-intercepted-enabled=true
