+++ quarkus-satoken-main/quarkus-satoken-resteasy/runtime/src/main/java/cn/dev33/satoken/annotation/SaCheckLogin.java
@@ -1,14 +1,17 @@
 package cn.dev33.satoken.annotation;
 
-import javax.interceptor.InterceptorBinding;
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
+import javax.interceptor.InterceptorBinding;
+
 /**
  * 登录认证：只有登录之后才能进入该方法
- * <p> 可标注在函数、类上（效果等同于标注在此类的所有方法上）
+ * <p>
+ * 可标注在函数、类上（效果等同于标注在此类的所有方法上）
+ *
  * @author kong
  *
  */
@@ -19,6 +22,7 @@ public @interface SaCheckLogin {
 
     /**
      * 多账号体系下所属的账号体系标识
+     *
      * @return see note
      */
     String type() default "";
