+++ quarkus-satoken-main/quarkus-satoken-resteasy/runtime/src/main/java/cn/dev33/satoken/annotation/SaCheckRole.java
@@ -1,39 +1,46 @@
 package cn.dev33.satoken.annotation;
 
-import javax.interceptor.InterceptorBinding;
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
+import javax.interceptor.InterceptorBinding;
+
 /**
- * 角色认证：必须具有指定角色标识才能进入该方法 
- * <p> 可标注在函数、类上（效果等同于标注在此类的所有方法上） 
+ * 角色认证：必须具有指定角色标识才能进入该方法
+ * <p>
+ * 可标注在函数、类上（效果等同于标注在此类的所有方法上）
+ *
  * @author kong
  *
  */
 @InterceptorBinding
 @Retention(RetentionPolicy.RUNTIME)
-@Target({ElementType.METHOD,ElementType.TYPE})
+@Target({ ElementType.METHOD, ElementType.TYPE })
 public @interface SaCheckRole {
 
-	/**
-	 * 需要校验的角色标识
-	 * @return 需要校验的角色标识
-	 */
-	String [] value() default {};
-
-	/**
-	 * 验证模式：AND | OR，默认AND
-	 * @return 验证模式
-	 */
-	SaMode mode() default SaMode.AND;
+    /**
+     * 需要校验的角色标识
+     *
+     * @return 需要校验的角色标识
+     */
+    String[] value() default {};
 
     /**
-     * 账号类型 
-     * <p> 建议使用常量，避免因错误拼写带来的bug 
-     * @return see note 
+     * 验证模式：AND | OR，默认AND
+     *
+     * @return 验证模式
      */
-	String type() default "";
-	
+    SaMode mode() default SaMode.AND;
+
+    /**
+     * 账号类型
+     * <p>
+     * 建议使用常量，避免因错误拼写带来的bug
+     *
+     * @return see note
+     */
+    String type() default "";
+
 }
