+++ quarkus-satoken-main/quarkus-satoken-resteasy/runtime/src/main/java/cn/dev33/satoken/annotation/SaCheckSafe.java
@@ -1,15 +1,18 @@
 package cn.dev33.satoken.annotation;
 
-import javax.interceptor.InterceptorBinding;
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
+import javax.interceptor.InterceptorBinding;
+
 /**
- * 二级认证校验：必须二级认证之后才能进入该方法 
- * 
- * <p> 可标注在函数、类上（效果等同于标注在此类的所有方法上） 
+ * 二级认证校验：必须二级认证之后才能进入该方法
+ *
+ * <p>
+ * 可标注在函数、类上（效果等同于标注在此类的所有方法上）
+ *
  * @author kong
  *
  */
@@ -19,9 +22,10 @@ import java.lang.annotation.Target;
 public @interface SaCheckSafe {
 
     /**
-     * 多账号体系下所属的账号体系标识 
-     * @return see note 
+     * 多账号体系下所属的账号体系标识
+     *
+     * @return see note
      */
-	String type() default "";
+    String type() default "";
 
 }
