+++ quarkus-satoken-main/quarkus-satoken-resteasy/runtime/src/main/java/io/quarkiverse/satoken/core/interceptor/AbstractSaInterceptor.java
@@ -1,25 +1,16 @@
 package io.quarkiverse.satoken.core.interceptor;
 
+import java.lang.annotation.Annotation;
 import java.lang.reflect.Method;
 import java.util.Objects;
 
-import javax.interceptor.AroundInvoke;
-import javax.interceptor.Interceptor;
 import javax.interceptor.InvocationContext;
 
-import cn.dev33.satoken.annotation.SaCheckBasic;
-import cn.dev33.satoken.annotation.SaCheckDisable;
-import cn.dev33.satoken.annotation.SaCheckLogin;
-import cn.dev33.satoken.annotation.SaCheckPermission;
-import cn.dev33.satoken.annotation.SaCheckRole;
-import cn.dev33.satoken.annotation.SaCheckSafe;
 import cn.dev33.satoken.annotation.SaIgnore;
 import cn.dev33.satoken.exception.BackResultException;
 import cn.dev33.satoken.exception.StopMatchException;
 import cn.dev33.satoken.fun.SaParamFunction;
 import cn.dev33.satoken.strategy.SaStrategy;
-import cn.dev33.satoken.util.SaTokenConsts;
-import io.quarkus.arc.Priority;
 
 /**
  * Sa-Token 综合拦截器，提供注解鉴权和路由拦截鉴权能力
@@ -27,22 +18,15 @@ import io.quarkus.arc.Priority;
  * @author nayan
  * @date 2022/4/14 4:17 PM
  */
-@SaCheckLogin
-@SaCheckRole
-@SaCheckPermission
-@SaCheckSafe
-@SaCheckDisable
-@SaCheckBasic
-@Interceptor
-@Priority(SaTokenConsts.ASSEMBLY_ORDER)
-public class SaInterceptor {
+
+public class AbstractSaInterceptor<T extends Annotation> {
 
     /**
      * 认证函数：每次请求执行
      * <p>
      * 参数：路由处理函数指针
      */
-    public SaParamFunction<Object> auth = handler -> {
+    public static SaParamFunction<Object> auth = handler -> {
     };
 
     /**
@@ -51,13 +35,11 @@ public class SaInterceptor {
      * @param auth /
      * @return 对象自身
      */
-    public SaInterceptor setAuth(SaParamFunction<Object> auth) {
-        this.auth = auth;
-        return this;
+    public static void setAuth(SaParamFunction<Object> auth) {
+        AbstractSaInterceptor.auth = auth;
     }
 
-    @AroundInvoke
-    public Object intercept(InvocationContext context) throws Exception {
+    public Object interceptAnnotation(InvocationContext context) throws Exception {
 
         try {
             // 获取此请求对应的 Method 处理函数
